+++ locales/MandarinTW.lua
@@ -1480,4 +1480,11 @@ GRML.MandarinTW = function()
     GRM_L["Due to limitations with macros a player can only move 1 rank at a time."] = true
     GRM_L["Disabled at Current Rank"] = true
     
+    -- 1.922
+    GRM_L["Verifying Macro Changes. One moment."] = true
+    GRM_L["Not all macro changes validated. One moment..."] = true
+    GRM_L["Macro rank changes have been validated!"] = true
+    GRM_L["Warning! Macro changes were not able to be validated. Please verify expected results before using the macro tool further."] = true
+
+    
 end